r10357/sofa : miniFlowVR: fixing gcc4.7 warnings
Former-commit-id: f148e106f771e01df0ad2beb13b4d190ac7b7314

+++ extlibs/miniFlowVR/src/ftlm/type.cpp
@@ -314,22 +314,22 @@ typedef struct {
 
 TypeToString names[] =
 { 
-  { Delete, "Delete" },
-  { Null, "Null" },
-  { Enum, "Enum" },
-  { ID, "ID" },
-  { Bool, "Bool" },
-  { Byte, "Byte" },
-  { Short, "Short" },
-  { Int, "Int" },
-  { Long, "Long" },
-  { QWord, "QWord" },
-  { Half, "Half" },
-  { Float, "Float" },
-  { Double, "Double" },
-  { LongDouble, "LongDouble" },
-  { String0, "String" },
-  { Null, NULL }
+  { (int)Delete, "Delete" },
+  { (int)Null, "Null" },
+  { (int)Enum, "Enum" },
+  { (int)ID, "ID" },
+  { (int)Bool, "Bool" },
+  { (int)Byte, "Byte" },
+  { (int)Short, "Short" },
+  { (int)Int, "Int" },
+  { (int)Long, "Long" },
+  { (int)QWord, "QWord" },
+  { (int)Half, "Half" },
+  { (int)Float, "Float" },
+  { (int)Double, "Double" },
+  { (int)LongDouble, "LongDouble" },
+  { (int)String0, "String" },
+  { (int)Null, NULL }
 };
 
 std::string name(int t)
